+++ @77bn/Addons/77_Armor/Data/Helmet/77_odst_h_helmet.rvmat
@@ -13,7 +13,7 @@ PixelShaderID = "Super";
 VertexShaderID = "Super";
 class Stage1
 {
-	texture = "77_Armor\Data\Helmet\77_odst_helmet_nohq.paa";
+	texture = "77_Armor\Data\Helmet\77_odst_h_helmet_nohq.paa";
 	uvSource = "tex";
 	class uvTransform
 	{
@@ -49,7 +49,7 @@ class Stage3
 };
 class Stage4
 {
-	texture = "77_Armor\Data\Helmet\77_odst_helmet_as.paa";
+	texture = "77_Armor\Data\Helmet\77_odst_h_helmet_as.paa";
 	uvSource = "tex";
 	class uvTransform
 	{
@@ -61,7 +61,7 @@ class Stage4
 };
 class Stage5
 {
-	texture = "77_Armor\Data\Helmet\77_odst_helmet_smdi.paa";
+	texture = "77_Armor\Data\Helmet\77_odst_h_helmet_smdi.paa";
 	uvSource = "tex";
 	class uvTransform
 	{
